+++ src/scenes/BootScene.js
@@ -7,6 +7,14 @@ export default class BootScene extends Phaser.Scene {
 
   preload() {
     this.load.image('logo', 'src/assets/logo.png');
+    // map tiles
+    this.load.image('tiles', 'src/assets/map/spritesheet.png');
+
+    // map in json format
+    this.load.tilemapTiledJSON('map', 'src/assets/map/map.json');
+
+    // our two characters
+    this.load.spritesheet('player', 'src/assets/RPG_assets.png', { frameWidth: 16, frameHeight: 16 });
   }
 
   create() {
